requirements.txt
llama-index
llama-index-core
llama-parse
openai
llama-index-llms-openai
python-dotenv
giskard[llm]
IPython
langchain-core
ragas
pyarrow
